add visitor for deprecated TorchVision transform
