+++ tools/torchfix/tests/fixtures/vision/checker/to_tensor.py
@@ -1,0 +1,26 @@
+from torchvision.transforms.v2 import ToTensor
+
+from torchvision.transforms.v2 import ToTensor as ToTensorAlias
+
+from torchvision.transforms.v2 import (
+    ToImage,
+    ToTensor,
+    ToDtype,
+)
+
+from torchvision.transforms import v2
+
+v2.ToTensor
+
+from torchvision.transforms import v2 as transforms
+
+transforms.ToTensor
+
+
+import torchvision.transforms.v2 as transforms
+
+transforms.ToTensor
+
+import torchvision.transforms.v2
+
+torchvision.transforms.v2.ToTensor
